Implement a rudimentary copy()
It's fairly rough, but it's also fairly correct as far as I can tell. I'm
not too happy with this code at present as far as "neatness" goes, due to
an abundance of copy-paste, but it's not incorrect or all _that_
suboptimal.

The tests verify that it all works, too.